app re-architecture
