+++ src/components/Header.js
@@ -8,7 +8,7 @@ const Header = ({ name, temp, desc }) => {
       { !!name ?
       <section id='current-forecast'>
         The current forecast for {name} is {desc} at {Math.round(temp)}°F
-        <Link to='/forecast' className='link'> Extended Forecast </Link>
+        {/* <Link to='/forecast' className='link'> Extended Forecast </Link> */}
       </section>
       : null }
     </article>
