+++ src/containers/App.js
@@ -1,7 +1,29 @@
+import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import App from '../components/app';
 import { fetchWeatherCurrentLocation, fetchWeatherPinnedLocation } from '../actions/index'
+import Header from '../containers/Header'
+
+class App extends Component {
+	componentWillMount() {
+		this.getCurrentLocation()
+	}
+
+	getCurrentLocation(){
+			 navigator.geolocation.getCurrentPosition((position) => {
+				 this.props.fetchWeatherCurrentLocation(position)
+			 })
+		 }
+
+	render() {
+		return (
+			<article>
+				<Header {...this.props}/>
+				{this.props.children}
+			</article>
+		)
+	}
+}
 
 const mapStateToProps = state => {
 	if (!state.LocalWeatherReducer.name) {
